face-recognition-system/requirement.txt
cmake==3.18.4.post1
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-Cors==3.0.10
numpy==1.20.2
opencv-python==4.5.2.52
requests==2.25.1
mysql-connector-python==8.0.24